Revert "Add component folder, format code"
This reverts commit 14e3c1fdba68774ec9799693489190ec1e99d730.
