+++ backend/package.json
@@ -1,35 +1,35 @@
 {
-    "name": "backend",
-    "version": "1.0.0",
-    "description": "",
-    "main": "./src/server.js",
-    "scripts": {
-        "start": "node -r esm ./src/server.js",
-        "server": "npm run start",
-        "production": "cross-env NODE_ENV=production PORT=3000 node -r esm ./src/server.js",
-        "init-db": "node -r esm ./src/db/init-db.js",
-        "test": "jest"
-    },
-    "keywords": [],
-    "author": "",
-    "license": "ISC",
-    "dependencies": {
-        "axios": "^0.21.1",
-        "dayjs": "^1.10.4",
-        "dummy-json": "^3.0.1",
-        "esm": "^3.2.25",
-        "express": "^4.17.1",
-        "mongoose": "^5.12.0",
-        "multer": "^1.4.2",
-        "uuid": "^8.3.2"
-    },
-    "devDependencies": {
-        "@babel/core": "^7.13.10",
-        "@babel/preset-env": "^7.13.10",
-        "@types/jest": "^26.0.21",
-        "babel-jest": "^26.6.3",
-        "cross-env": "^7.0.3",
-        "jest": "^26.6.3",
-        "mongodb-memory-server": "^6.9.6"
-    }
+  "name": "backend",
+  "version": "1.0.0",
+  "description": "",
+  "main": "./src/server.js",
+  "scripts": {
+    "start": "node -r esm ./src/server.js",
+    "server": "npm run start",
+    "production": "cross-env NODE_ENV=production PORT=3000 node -r esm ./src/server.js",
+    "init-db": "node -r esm ./src/db/init-db.js",
+    "test": "jest"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "dependencies": {
+    "axios": "^0.21.1",
+    "dayjs": "^1.10.4",
+    "dummy-json": "^3.0.1",
+    "esm": "^3.2.25",
+    "express": "^4.17.1",
+    "mongoose": "^5.12.0",
+    "multer": "^1.4.2",
+    "uuid": "^8.3.2"
+  },
+  "devDependencies": {
+    "@babel/core": "^7.13.10",
+    "@babel/preset-env": "^7.13.10",
+    "@types/jest": "^26.0.21",
+    "babel-jest": "^26.6.3",
+    "cross-env": "^7.0.3",
+    "jest": "^26.6.3",
+    "mongodb-memory-server": "^6.9.6"
+  }
 }
