+++ frontend/src/HomePage.js
@@ -1,0 +1,13 @@
+import Footer from './Footer';
+import Body from './Body';
+export default function HomePage() {
+    return (
+        <>
+            <Body></Body>
+            <Footer></Footer>
+        </>
+
+    )
+
+
+}
