--- frontend/src/components/HomePage.js
@@ -1,11 +1,0 @@
-import Body from "./Body";
-import Footer from "./Footer";
-
-export default function HomePage() {
-    return (
-        <>
-            <Body/>
-            <Footer/>
-        </>
-    );
-}
